scraper/src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.0
    # via scraper (pyproject.toml)
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via httpx
attrs==25.4.0
    # via aiohttp
beautifulsoup4==4.14.2
    # via bs4
bs4==0.0.2
    # via scraper (pyproject.toml)
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.3
    # via requests
colorama==0.4.6
    # via tqdm
cryptography==46.0.2
    # via pyjwt
deprecation==2.1.0
    # via
    #   postgrest
    #   storage3
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   postgrest
    #   storage3
    #   supabase
    #   supabase-auth
    #   supabase-functions
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
packaging==25.0
    # via deprecation
postgrest==2.22.0
    # via supabase
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
psycopg2==2.9.11
    # via scraper (pyproject.toml)
pycparser==2.23
    # via cffi
pydantic==2.12.0
    # via
    #   postgrest
    #   realtime
    #   storage3
    #   supabase-auth
pydantic-core==2.41.1
    # via pydantic
pyjwt==2.10.1
    # via supabase-auth
python-dotenv==1.1.1
    # via scraper (pyproject.toml)
realtime==2.22.0
    # via supabase
requests==2.32.5
    # via scraper (pyproject.toml)
sniffio==1.3.1
    # via anyio
soupsieve==2.8
    # via beautifulsoup4
sqlalchemy==2.0.44
    # via scraper (pyproject.toml)
storage3==2.22.0
    # via supabase
strenum==0.4.15
    # via supabase-functions
supabase==2.22.0
    # via scraper (pyproject.toml)
supabase-auth==2.22.0
    # via supabase
supabase-functions==2.22.0
    # via supabase
tqdm==4.67.1
    # via scraper (pyproject.toml)
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   beautifulsoup4
    #   pydantic
    #   pydantic-core
    #   realtime
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via scraper (pyproject.toml)
urllib3==2.5.0
    # via requests
websockets==15.0.1
    # via realtime
yarl==1.22.0
    # via
    #   aiohttp
    #   postgrest
    #   storage3
    #   supabase-functions
